backend/requirements.txt
# FastAPI and core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database and ORM
sqlalchemy>=2.0.25
alembic>=1.13.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data processing - simplified for Python 3.13 compatibility
pandas>=2.2.0

# CORS and middleware
fastapi-cors>=0.0.6

# Environment variables
python-dotenv>=1.0.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
